Build model-v2 inference foundation
Add the portable engine interfaces, streaming model-v2 tooling and round-trip tests. Isolate deterministic v1 fixtures, fail closed for unsupported production inference, and align CI, QEMU contracts, documentation, and readiness reporting with verified capabilities.

+++ userspace/lib/xaios_user.c
@@ -170,7 +170,7 @@ int xaios_cpu_ai_decode(const void *input, u64 input_size, char *output,
   request.out_size = (u64)out_size;
   u64 rc = xaios_syscall3(XAIOS_SYSCALL_CPU_AI_DECODE, (u64)&request,
                          sizeof(request), 0);
-  return rc == ~0ULL ? -1 : (int)rc;
+  return rc == ~0ULL ? -1 : (int)(signed long long)rc;
 }
 
 int xaios_remote_login(const char *user, const char *command, char *output,
